sdk: move setAttachedPreimage from TestPreimages to TestLookup
Aligns the AS-side class boundary with the JS-side file boundary
after Task 8's refactor moved the attached-preimage map from
accumulate/preimages.ts to general/lookup.ts.

- sdk/test/test-ecalli/lookup.ts: owns setAttachedPreimage +
  clearAttachedPreimages now (plus the two @external declarations).
  Imports Bytes32/BytesBlob since it needs them.
- sdk/test/test-ecalli/preimages.ts: trimmed back to the four
  accumulate-context stubs (query/solicit/forget/provide). No
  Bytes32/BytesBlob import needed anymore.
- examples/pastebin: single call-site migration
  (TestPreimages.setAttachedPreimage → TestLookup.setAttachedPreimage).
- docs/src/testing.md: moved the "Simulating extrinsic-driven preimage
  delivery" subsection from under TestPreimages to under TestLookup.
- sdk-ecalli-mocks: cross-reference comment updated.

Tests still 6/6 pastebin + 207/207 SDK.
